@@ -96,7 +96,7 @@ public enum EDatabaseComponentName implements IComponentName {
     JAVADB(DatabaseConnectionItem.class, EDatabaseTypeName.JAVADB, "tJavaDBInput", "tJavaDBOutput", true), //$NON-NLS-1$ //$NON-NLS-2$
     INGRES(DatabaseConnectionItem.class, EDatabaseTypeName.INGRES, "tIngresInput", "tIngresOutput", true), // "INGRES"), //$NON-NLS-1$ //$NON-NLS-2$
     INTERBASE(DatabaseConnectionItem.class, EDatabaseTypeName.INTERBASE, "tInterbaseInput", "tInterbaseOutput", true), // "INTERBASE" //$NON-NLS-1$ //$NON-NLS-2$
-    EXASOL(DatabaseConnectionItem.class, EDatabaseTypeName.EXASOL, "tEXAInput", "tEXAOutput", true), // "INGRES"), //$NON-NLS-1$ //$NON-NLS-2$
+    EXASOL(DatabaseConnectionItem.class, EDatabaseTypeName.EXASOL, "tExasolInput", "tExasolOutput", true), // "INGRES"), //$NON-NLS-1$ //$NON-NLS-2$
     VECTORWISE(DatabaseConnectionItem.class, EDatabaseTypeName.VECTORWISE, "tVectorWiseInput", "tVectorWiseOutput", true), //$NON-NLS-1$ //$NON-NLS-2$
     // )
     // ,
